Made a buncha changes related to 64-bit compatibility or performance, tidied up some things and implemented some NPCs

+++ source/player.h
@@ -67,7 +67,7 @@ class player {
 		void setPos(int setX, int setY);
 		void setDir(int setDirect);
 
-		void damage(int damage);
+		void damage(int16_t damage);
 
 		void actNormal(bool bKey);
 		void actStream(bool bKey);
